Added predictions metadata concept to server. Renamed existing predictions route to inference.
