+++ api/routes/solutions.go
@@ -45,15 +45,8 @@ type SolutionResponse struct {
 func SolutionsHandler(solutionCtor model.SolutionStorageCtor) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// extract route parameters
-		dataset := pat.Param(r, "dataset")
-		target := pat.Param(r, "target")
-
-		if dataset == "null" {
-			dataset = ""
-		}
-		if target == "null" {
-			target = ""
-		}
+		dataset := handleNullParameter(pat.Param(r, "dataset"))
+		target := handleNullParameter(pat.Param(r, "target"))
 
 		solution, err := solutionCtor()
 		if err != nil {
