+++ main.go
@@ -227,10 +227,10 @@ func main() {
 	registerRoute(mux, "/distil/solutions/:dataset/:target", routes.SolutionsHandler(pgSolutionStorageCtor))
 	registerRoute(mux, "/distil/solution-requests/:dataset/:target", routes.SolutionRequestsHandler(pgSolutionStorageCtor))
 	registerRoute(mux, "/distil/solution-request/:request-id", routes.SolutionRequestHandler(pgSolutionStorageCtor))
-	registerRoute(mux, "/distil/prediction/:prediction-id", routes.SolutionHandler(pgSolutionStorageCtor))
-	registerRoute(mux, "/distil/predictions/:fitted-solution-id", routes.SolutionsHandler(pgSolutionStorageCtor))
-	registerRoute(mux, "/distil/prediction-requests/:fitted-solution-id", routes.SolutionRequestsHandler(pgSolutionStorageCtor))
-	registerRoute(mux, "/distil/prediction-requests/:request-id", routes.SolutionRequestHandler(pgSolutionStorageCtor))
+	registerRoute(mux, "/distil/prediction/:prediction-id", routes.PredictionHandler(pgSolutionStorageCtor))
+	registerRoute(mux, "/distil/predictions/:fitted-solution-id", routes.PredictionsHandler(pgSolutionStorageCtor))
+	registerRoute(mux, "/distil/prediction-requests/:fitted-solution-id", routes.PredictionRequestsHandler(pgSolutionStorageCtor))
+	registerRoute(mux, "/distil/prediction-requests/:request-id", routes.PredictionRequestHandler(pgSolutionStorageCtor))
 	registerRoute(mux, "/distil/variables/:dataset", routes.VariablesHandler(esMetadataStorageCtor, pgDataStorageCtor))
 	registerRoute(mux, "/distil/variable-rankings/:dataset/:target", routes.VariableRankingHandler(esMetadataStorageCtor, pgSolutionStorageCtor, pgDataStorageCtor))
 	registerRoute(mux, "/distil/residuals-extrema/:dataset/:target", routes.ResidualsExtremaHandler(esMetadataStorageCtor, pgSolutionStorageCtor, pgDataStorageCtor))
@@ -257,7 +257,7 @@ func main() {
 	registerRoutePost(mux, "/distil/geocode/:dataset/:variable", routes.GeocodingHandler(esMetadataStorageCtor, pgDataStorageCtor))
 	registerRoutePost(mux, "/distil/cluster/:dataset/:variable", routes.ClusteringHandler(esMetadataStorageCtor, pgDataStorageCtor))
 	registerRoutePost(mux, "/distil/upload/:dataset", routes.UploadHandler(path.Join(config.D3MOutputDir, config.AugmentedSubFolder), ingestConfig))
-	registerRoutePost(mux, "/distil/predict/:dataset/:target-type/:fitted-solution-id", routes.PredictionsHandler(path.Join(config.D3MOutputDir, config.AugmentedSubFolder), pgDataStorageCtor, pgSolutionStorageCtor, esMetadataStorageCtor, &config, ingestConfig))
+	registerRoutePost(mux, "/distil/predict/:dataset/:target-type/:fitted-solution-id", routes.InferenceHandler(path.Join(config.D3MOutputDir, config.AugmentedSubFolder), pgDataStorageCtor, pgSolutionStorageCtor, esMetadataStorageCtor, &config, ingestConfig))
 	registerRoutePost(mux, "/distil/join", routes.JoinHandler(esMetadataStorageCtor))
 	registerRoutePost(mux, "/distil/timeseries/:dataset/:timeseriesColName/:xColName/:yColName/:timeseriesURI/:invert", routes.TimeseriesHandler(pgDataStorageCtor))
 	registerRoutePost(mux, "/distil/timeseries-forecast/:dataset/:timeseriesColName/:xColName/:yColName/:timeseriesURI/:result-uuid", routes.TimeseriesForecastHandler(pgDataStorageCtor, pgSolutionStorageCtor))
